Done! Everything works just fine! All warnings fixed. Code optimized...

+++ src/utility/Utility.java
@@ -8,7 +8,7 @@ public class Utility {
 
     public static int charToAsciiCode(char caracter)
     {
-        return (int) caracter;
+        return  caracter;
     }
 
 }
